# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Flyspray",
        "version": "1.0.0",
        "contact": {}
    },
    "paths": {
...
        "/component": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new component for a project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Components"
                ],
                "summary": "Create a component",
                "parameters": [
                    {
                        "description": "Component data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.createComponentInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "component is created successfully (Component details in the 'Data' field)",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "Failed to read component data",
                        "schema": {}
                    },
                    "401": {
                        "description": "Authentication is required",
                        "schema": {}
                    },
                    "403": {
                        "description": "Do not have access to create component",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
        "/component/filters": {
            "get": {
                "description": "Get a list of components based on filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Components"
                ],
                "summary": "Get components",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Component name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Creation date (after)",
                        "name": "after",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Components are retrieved successfully (Components details in the 'Data' field)",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseMsg"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
        "/component/{id}": {
            "get": {
                "description": "Get a component by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Components"
                ],
                "summary": "Get a component",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Component ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "component is retrieved successfully (Component details in the 'Data' field)",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseMsg"
                        }
                    },
                    "404": {
                        "description": "Component is not found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update an existing component",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Components"
                ],
                "summary": "Update a component",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Component ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Component data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.updateComponentInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "component is updated successfully",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "Failed to read component data",
                        "schema": {}
                    },
                    "401": {
                        "description": "Authentication is required",
                        "schema": {}
                    },
                    "403": {
                        "description": "Do not have access to update component",
                        "schema": {}
                    },
                    "404": {
                        "description": "Component is not found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a component by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Components"
                ],
                "summary": "Delete a component",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Component ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Component is deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/app.ResponseMsg"
                        }
                    },
                    "401": {
                        "description": "Authentication is required",
                        "schema": {}
                    },
                    "403": {
                        "description": "Do not have access to delete component",
                        "schema": {}
                    },
                    "404": {
                        "description": "Component is not found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
        "app.ResponseMsg": {
            "type": "object",
            "properties": {
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "app.createComponentInput": {
            "type": "object",
            "required": [
                "name",
                "project_id"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "project_id": {
                    "type": "string"
                }
            }
        },
...
        "app.updateComponentInput": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}